removes unneeded file and moves format_time to a more global location
