--- src/dw_design_system/dwds/components/cta_card.scss
@@ -1,1 +1,0 @@
-.dwds-card.dwds-cta-card {}
