+++ src/dw_design_system/dwds/stylesheet.scss
@@ -10,7 +10,6 @@
 // Components
 @forward "components/banner_card";
 @forward "components/card_group";
-@forward "components/cta_card";
 @forward "components/engagement_card";
 @forward "components/link_list";
 @forward "components/navigation_card";
