+++ src/events/utils.py
@@ -1,20 +1,5 @@
-from datetime import time
-
+from core.utils import format_time
 from events.models import EventPage
-
-
-def format_time(time_obj: time) -> str:
-    """
-    Build a time string that shows the needed information.
-
-    Returns:
-     - If there are minutes on the hour: `10:30` -> `10:30am`
-     - If there are not minutes on the hour: `10:00` -> `10am`
-
-    """
-    if time_obj.minute == 0:
-        return time_obj.strftime("%-I%p").lstrip("0").lower()
-    return time_obj.strftime("%-I:%M%p").lstrip("0").lower()
 
 
 def get_event_date(event: EventPage) -> str:
